requirements.txt
feedparser
slackclient
beautifulsoup4